restructure code, extensible manifest
